Add CodeBuild support

+++ lintly/builds.py
@@ -1,8 +1,8 @@
 import logging
 
 from .exceptions import NotPullRequestException
-from .git.github import GitHubBackend
-from .git.errors import GitClientError, NotFoundError
+from .backends.github import GitHubBackend
+from .backends.errors import GitClientError, NotFoundError
 from .formatters import build_pr_comment
 from .parsers import PARSERS
 from .projects import Project
@@ -45,34 +45,37 @@ class LintlyBuild(object):
         """
         Posts a comment to the GitHub PR if the diff results have issues.
         """
-        post_pr_comment = True
+        if self.has_violations:
+            post_pr_comment = True
 
-        # Attempt to post a PR review. If posting the PR review fails because the bot account
-        # does not have permission to review the PR then simply revert to posting a regular PR
-        # comment.
-        try:
-            logger.info('Deleting old PR review comments')
-            self.git_client.delete_pull_request_review_comments(self.config.pr)
+            # Attempt to post a PR review. If posting the PR review fails because the bot account
+            # does not have permission to review the PR then simply revert to posting a regular PR
+            # comment.
+            try:
+                logger.info('Deleting old PR review comments')
+                self.git_client.delete_pull_request_review_comments(self.config.pr)
 
-            logger.info('Creating PR review')
-            self.git_client.create_pull_request_review(self.config.pr, self.all_violations)
-            post_pr_comment = False
-        except GitClientError as e:
-            # TODO: Make `create_pull_request_review` raise an `UnauthorizedError`
-            # so that we don't have to check for a specific message in the exception
-            if 'Viewer does not have permission to review this pull request' in str(e):
-                logger.info("Could not post PR review (the bot account didn't have permission)")
-                pass
-            else:
-                raise
+                logger.info('Creating PR review for PR #{}'.format(self.config.pr))
 
-        if post_pr_comment:
-            # logger.info('Deleting old PR comment')
-            # self.git_client.delete_pull_request_comments(self.config.pr, bot=self.bot.username)
+                # Pull request reviews must always run against the diff violations
+                self.git_client.create_pull_request_review(self.config.pr, self._diff_violations)
+                post_pr_comment = False
+            except GitClientError as e:
+                # TODO: Make `create_pull_request_review` raise an `UnauthorizedError`
+                # so that we don't have to check for a specific message in the exception
+                if 'Viewer does not have permission to review this pull request' in str(e):
+                    logger.info("Could not post PR review (the bot account didn't have permission)")
+                    pass
+                else:
+                    raise
 
-            logger.info('Creating PR comment for')
-            comment = build_pr_comment(self.config, self.all_violations)
-            self.git_client.create_pull_request_comment(self.config.pr, comment)
+            if post_pr_comment:
+                # logger.info('Deleting old PR comment')
+                # self.git_client.delete_pull_request_comments(self.config.pr, bot=self.bot.username)
+
+                logger.info('Creating PR comment for')
+                comment = build_pr_comment(self.config, self.violations)
+                self.git_client.create_pull_request_comment(self.config.pr, comment)
 
     def post_commit_status(self):
         """
